Creates a proper implementation of merge sort
The merge() function was originally having harsh indexing issues where
the function was accessing values outside the scope of the arrays it was
meant to develop on. This has been fixed. The merge() function also makes
sure to check if we've already sorted through one of the halves passed
into the function.
Fixes a problem in the mergeSort() function which was originally
accounting for the fact that the end of array index parameter, b, is
meant to represent the final index, rather than the length of the entire
array. The difference is one off, but would confuse the algorithm
regardless.
Adds citation of the source I learned about this algorithm.