First implementation of max subarray problem
Adds the first implementation to find the max subarray in an array of
integers.
maxCrossSub() finds the subarray within the array arr[] that both
passes the midpoint of the array, and also has the largest sum of all
the numbers in the subarray.
maxSub() takes the subarray on the left and right halves of the array
arr[] and then recursively calls itself on those halves, as well as
calling maxCrossSub() in the process in order to eventually find the
subarray with the largest sum.