Modifies 3Lab code to function properly
Changes the maxSub() and maxCrossSub() methods in awoersching.cpp to
instead return an integer value of the sum of the subarray that's being
analyzed. This change was made after reading the lab instructions more
carefully and realizing that there was no need to save the bounds of
these subarrays. This made a lot of code unnecessary, so many changes
were made to adjust to the new return type.
Also attempted to edit the .gitignore to remove the a.exe and result
files from my commits.