Implements quick sort algorithm
Creates the quick sort algorithm, has NOT been debugged yet.
randoSplit() is meant to randomly partition the array and return the
index where the array is being partitioned.
randoQuick() is meant to be the main recursive sorting the algorithm.
part() is meant to partition the array by choosing the pivot point,
which has been randomized by randoSplit() and then sorting the array
from the beginning to the pivot point based one whether or not an
element is smaller than the element at the pivot point.
swap() is simply meant to swap two elements in an array.