Creates method to make array into max heap array
While creating the heap sort algorithm, it's necessary to make sure the
array is sorted in a manner that would suggest that it has max heap
properties. Thus the max-heap() method is created.
The left(), right(), and parent() methods are simple helper methods
meant to return the index of the left, right, and parent node of a
binary tree respectively.
swap() is a helper function taken from the quick sort algorithm that
switches any two elements in an array.