Creates functioning heap sort algorithm
Debugs code to create a properly functioning heap sort algorithm.
Fixes if statements in maxHeap() that was causing the method to make
comparisons that incorrectly sorted the array.
Fixes heapSort() function by changing the starting index of the second
for loop. This was causing the for loop to start outside of the bounds
of the array. Also changes the size parameter when calling maxHeap()
because the heap size is meant to decrease as heapSort() continues.