Creates functioning Radix Sort
Has a lot of changes.
Insertion sort was a bad idea. Using an O(n^2) sorting algorithm will
have issues with the radix sorting algorithm which requires O(n).
Insertion sort was completely replaced with counting sort.
countSort() takes in all vector arrays and the index for which we are
sorting and sorts based on that index by counting how many of each
number there is and keeping track using the array counts[]. Index 0
counts the number of zeros being sorted, while index 1 counts the number
of ones being sorted and so on. Then from there we are able to use these
numbers to figure out which indicies in the sorted[] array they'll need
to be in.
After much debugging an issue was revealed that the vector array being
sorted (arr[]) was always meant to be kept up to date with each
iteration of the for loop in the radix sort function radSort(), this
seems like a simple fix but was met with many seg faults. So in order to
combat this issue, the radSort() method checks if the index being sorted
upon is an even or odd number and then will either sort arr[] or
sorted[] and store the newly sorted version in the opposite vector
array. This can also be seen in main() and prints the final product
based on the lengths of the vectors (which in this case is always 10).