Creates some necessary helper functions
This creates three basic functions related to hash tables.
h() acts as the simple hash key. Takes in the key, k, and returns the
bucket this number, k, should be in.
i() takes in a key, k, and inserts the number at the beginning of the
queue it belongs in.
d() deletes the key, k, from the hash table assuming it's there in the
first place. If it is there, the function will print that the number, k,
has been deleted, and will specify otherwise if the key, k, was not
found.