Creates remaining helper functions
s() searches for an element, k, in the hash table. Prints the location
if k is found in the hash table and prints ":NOT_FOUND" otherwise.
o() prints the entire hash table.
Also makes variable m a global variable, which makes it easier to
implement these helper functions. Considering doing the same with other
variables.