Implements functioning hash table
Debugs helper functions.
Creates new helper function strToInt() which takes in a string and
returns an integer. Created to help call the functions in main().
main() now asks the user for input and parses it to check which command
the user wants to do. For example, the command "i2" will insert the
number 2 into the correct bucket of the hash table. "d2" will delete
that number from the hash table and output that the value was deleted.
"o" will simply output the entire table, and "s2" will search for the
first instance of the number 2 in the hash table and output whether or
not the value exists and where it exists.
These commands work for any n digit number.