Modified cutRod() to print the lengths of rods
Added a slices[] array and a while loop that keeps track of, and prints
the correct number of slices of a rod.
Changes cutRod() to a void function, as it only needs to print the
necessary values, rather than store them in memory.
Restores verifier.cpp to it's original state, as it was not clear what
verifier.cpp was meant for originally. verifier.cpp is meant to help
Grade assess awoersching.cpp for it's proper functionality.