Starts on the course to find optimal matrix mult
Creates two functions:
matChain() takes in the array of matrix dimensions, arr[], and the 2D
array s which will be manipulated by matChain() to hold the optimal
indices from which we can split arr in order to do the lowest amount of
operations.
printMatrix() as it states, it takes in a matrix and its dimensions, and
prints said matrix.