Redesigns approach to find lowest number of mults
Redesigns matChain() to use memorization, such that the algorithm should
have a slightly lower run time compared to the previous approach.
lookup() recursively looks through the 2D array m[][] in order to more
easily find the lowest number of multiplication processes needed to
multiply all matricies together.
printOptimal() is an attempt at finding the optimal order to multiply
the matricies together. However, more work is needed in order to create
a functional input, s.
This commit correctly displays the lowest number of multiplication
processes necessary, however, still does not print the order in which
the matricies need to be multiplied in order to resemble that same
number.