Properly implements Matrix Chain Multiplication
Adds a 2D array, s, initialized within matChain() meant to hold the
optimal indices to split the array, arr, to find the most optimal way
to multiply each matrix together.
lookup() was slightly changed to take in this new array, s, and modify
it properly. This made the printOptimal() method useful.
matChain() was changed to a void function, as it was redundant to have
it returning the value we were only meant to print.