Adds input structure and creates node struct
Creates a structure named node containing a frequency of a particular
letter appearing, as well as a left and right pointer to other nodes.
Edits main() to take in user input of the 6 different frequencies of
each letter and insert them into the array of nodes.
Begins work on greed() algorithm which will sort all 6 frequencies using
the Huffman code algorithm and print the final codes that will be used
to identify each of the 6 characters.
Adds citation.
Note: Due to the lab assignment, the code is going to constantly assume
there will be exactly 6 characters that will need to be analyzed.