Creates vector of Nodes and min() function
A vector of Nodes has been created to act as a queue. All other methods
have been modified to reflect this change.
greed() now takes in the vector q, and enacts the Huffman algorithm onto
it.
min() returns the index from which the lowest frequency node is in the
vector q.
This code still has issues maintaining the final tree.