Replaces struct and adds helper functions
Replaces the struct Vert() with two vector<int> arrays, g[] & t[],
where t[] is the meant to be the transpose of g[].
transpose() takes in two vector<int> arrays g[] & t[], and transposes
g[] by storing it in t[].
printArray() simply takes in an array and prints the contents of each
index. Might delete later.