Creates basic dfs functions
Builds Depth First Search methods that will be used in the Strong
Connected Components algorithm.
dfs() takes in the 2D array g[] and begins enacting dfsv() on every node
that hasn't been visited in the process of calling dfsv().
dfsv() is a recursive function that visits all verticies in the map and
stores the start and finish times of each vertex. If a preceeding vertex
hasn't been visited, the function will call itself only changing the
parameter of the vertex it's currently at.