Initializes necessary variables in main()
Creates matrix of weights meant to represent the weights connected
between two verticies. If there is no connection between two verticies,
-1 is placed in the indicies as a place holder. This is assuming there
will be no negatively weighed edges.
The graph, G[], will be an array of vectors that hold all the verticies
connected to any one vertex. So the edges of vertex 1 will be held in
G[1] and so on.