Creates first prim() method, adds minor bug fixes
Creates general idea of what prim() should do. Meant to create a minimum
spanning tree and return the parents of every vertex. Still needs to be
worked on, as the function causes seg faults.
search() is a simple helper function that checks a vector to see if an
integer value, v, is present.
weights[][] was changed to be a double pointer as it is easier to pass
into the method prim().