Creates first implementation of Bellman-Ford
bell() is meant to implement the Bellman-Ford algorithm and print the
shortest paths from the source (vertex 0) assuming there are no
negative weight cycles. If there are, the function should simply print
"FALSE" and end the cycle.
Currently the issue with this code is that it always prints "FALSE" even
when there no negative weight cycles.