Fixes bell() function and other problems in main()
Fixed issues with variable name assignments. Used the variable name "v"
far too often, causing the program to have odd settings, like having a
for loop go until INT32_MAX, rather than something simple like 10.
In bell(), gets rid of parent[] as it ended up being unnecessary for the
purposes of the lab.
In bell() decided to start using longs V and UW as it was presumed to
have overflow errors when the program has integers set to INT32_MAX.
In bell() fixed the assignment of distances. Where as, before the
program set dist[k] equal to dist[j] + weight[j][k], now the program
replaced every k variable with vert[j][k] as this the same as observing
vertex v from vertex u.
In bell() changed the final if statement meant to print the distances
from the source vertex from if(noCycle) to if(!noCycle) as this ended up
checking if there was a wasn't a negative weight cycle and printing
"FALSE" for it rather than the opposite. From what could be understood,
this caused the program to print a negative weight cycle if there was
one, and print nothing if there wasn't one.